Move routing e2e to it's own suite
Because routing starts an event loop in a goroutine and doesn't clean
it up, this test can't be run in the same suite as other tests. that
event loop will interfere with the other tests, which assume it is
not running in the background
