+++ test/e2e/routing/routing_e2e_suite_test.go
@@ -1,0 +1,27 @@
+package routing
+
+import (
+	"testing"
+
+	"github.com/onsi/gomega/gexec"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+)
+
+func TestE2e(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "E2E Suite")
+}
+
+var PathToUds string
+
+var _ = BeforeSuite(func() {
+	var err error
+	PathToUds, err = gexec.Build("github.com/solo-io/solo-projects/projects/discovery/cmd")
+	Expect(err).ShouldNot(HaveOccurred())
+})
+
+var _ = AfterSuite(func() {
+	gexec.CleanupBuildArtifacts()
+})
